move settings test to basic tests
